+++ package.json
@@ -1,12 +1,13 @@
 {
   "name": "acode-plugin-acodex",
-  "version": "3.1.6",
+  "version": "3.1.8",
   "description": "Terminal support for Acode",
   "main": "dist/main.js",
   "repository": "https://github.com/bajrangCoder/acode-plugin-terminal.git",
   "author": "Raunak Raj <bajrangcoders@gmail.com>",
   "license": "MIT",
   "dependencies": {
+    "@novnc/novnc": "^1.5.0",
     "@xterm/addon-attach": "0.10.0-beta.1",
     "@xterm/addon-fit": "0.9.0-beta.1",
     "@xterm/addon-image": "^0.8.0",
@@ -15,7 +16,8 @@
     "@xterm/addon-web-links": "0.10.0-beta.1",
     "@xterm/addon-webgl": "0.17.0-beta.1",
     "@xterm/xterm": "^5.5.0",
-    "html-tag-js": "^1.7.1"
+    "html-tag-js": "^1.7.1",
+    "ollama": "^0.5.9"
   },
   "devDependencies": {
     "esbuild": "^0.19.12",
