+++ plugin.json
@@ -2,7 +2,7 @@
   "id": "bajrangcoder.acodex",
   "name": "AcodeX - Terminal",
   "main": "dist/main.js",
-  "version": "3.1.7",
+  "version": "3.1.8",
   "readme": "readme.md",
   "icon": "icon.png",
   "files": [],
