+++ src/utils/constants.js
@@ -5,6 +5,7 @@ acode.addIcon("deepseek", "https://raw.githubusercontent.com/deepseek-ai/DeepSee
 acode.addIcon("chatgpt", "https://upload.wikimedia.org/wikipedia/commons/0/04/ChatGPT_logo.svg");
 acode.addIcon("gemini", "https://upload.wikimedia.org/wikipedia/commons/4/45/Gemini_language_model_logo.png");
 
+export const GUI_VIEWER = false;
 export const IMAGE_RENDERING = false;
 export const ALLOW_TRANSPRANCY = false;
 export const CURSOR_BLINK = true;
@@ -154,8 +155,8 @@ export const AVAILABLE_AI_MODELS = [
   ],
   [
     "local-llm",
-    "Local LLM",
+    "Local LLM(ollama)",
     "icon phone_android",
-    false
+    true
   ]
 ];
